Rename 'users' to 'user'
